+++ statistics-impl/src/main/java/com/dwarfeng/statistics/impl/handler/driver/FixedRateDriverProvider.java
@@ -122,10 +122,10 @@ public class FixedRateDriverProvider implements DriverProvider {
                     return;
                 }
 
-                LOGGER.debug("计划时间已到达, fixed rate 驱动器驱动 {} 记录设置执行动作...", statisticsSettingKey);
+                LOGGER.debug("计划时间已到达, fixed rate 驱动器驱动统计设置 {} 执行统计动作...", statisticsSettingKey);
                 context.execute(statisticsSettingKey);
             } catch (Exception e) {
-                LOGGER.warn("记录 {} 时出现异常, 放弃本次记录", statisticsSettingKey, e);
+                LOGGER.warn("驱动设置 {} 执行统计动作时出现异常, 放弃本次记录", statisticsSettingKey, e);
             } finally {
                 lock.unlock();
             }